split calendar into separate functions
