+++ src/js/calendar/addLeadingZero.js
@@ -1,0 +1,3 @@
+export default function addLeadingZero(value) {
+    return String(value).padStart(2, '0');
+  }
