+++ src/js/calendar/onCalendarFormClick.js
@@ -1,0 +1,6 @@
+import { ref } from './refs-calendar';
+
+export default function onCalendarFormClick() { 
+  ref.calendarWrapper.classList.toggle('hidden');
+  ref.calendarForm.classList.toggle('selected');
+}
