+++ src/js/calendar/refs-calendar.js
@@ -1,0 +1,8 @@
+export const ref = {
+  currentMonthYear: document.querySelector('.calendar__title'),
+  days: document.querySelector('.days'),
+  prevNextArrows: document.querySelector('.calendar__arrows'),
+  calendarWrapper: document.querySelector('.calendar__wrapper'),
+  calendarForm: document.querySelector('.calendar__form'),
+  calendarDisplayDate: document.querySelector('.calendar__form-text'),
+};
